Move all EnvironmentHelper tests into new file
